move apiConsumerRoute in adminRoutes
