+++ shared/src/index.ts
@@ -10,8 +10,7 @@ export * from "./agency/agency.dto";
 export * from "./agency/agency.schema";
 export * from "./agency/AgencyDtoBuilder";
 export * from "./apiConsumer/ApiConsumer";
-export * from "./apiConsumer/apiConsumer.routes";
-export * from "./apiConsumer/ApiConsumer.schema";
+export * from "./apiConsumer/apiConsumer.schema";
 export * from "./Builder";
 export * from "./contactEstablishmentRequest/contactEstablishmentRequest.dto";
 export * from "./contactEstablishmentRequest/contactEstablishmentRequest.schema";
